Fix Raw SQL bug to prevent duplicate siteAlerts creation
 "SiteAlert"."isProcessed" = false
had made this raw SQL buggy as the code could not find duplicates since the new siteAlerts would always have isProcessed as false, and the old one which have already been notified (which is automated) will always be true, so each cron constantly created the same siteAlert with each call, and notified the siteAlert constantly, and then became isProcessed to false. This had created a loop of always making each siteAlert being unique despite not being so.